webapp/requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-Session==0.5.0
python-ldap==3.4.3
requests==2.31.0
gunicorn==21.2.0
redis==5.0.1
Werkzeug==3.0.1
